requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
boto3==1.38.32
botocore==1.38.32
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.18.0
grpcio==1.67.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.64
langchain-openai==0.3.21
langsmith==0.3.45
milvus-lite==2.4.12
numpy==2.2.6
openai==1.84.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
platformdirs==4.3.8
protobuf==6.31.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pymilvus==2.5.10
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
tenacity==9.1.2
tiktoken==0.9.0
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
zstandard==0.23.0
